Add wasmJs downloader
