+++ platformtools/releasefetcher/src/androidJvmMain/kotlin/io/github/kdroidfilter/platformtools/releasefetcher/downloader/PlatformFile.androidJvm.kt
@@ -1,0 +1,3 @@
+package io.github.kdroidfilter.platformtools.releasefetcher.downloader
+
+actual typealias PlatformFile = java.io.File
