+++ platformtools/releasefetcher/src/androidJvmMain/kotlin/io/github/kdroidfilter/platformtools/releasefetcher/github/VersionUtils.androidJvm.kt
@@ -1,0 +1,5 @@
+package io.github.kdroidfilter.platformtools.releasefetcher.github
+
+import io.github.kdroidfilter.platformtools.getAppVersion
+
+actual fun getCurrentAppVersion(): String = getAppVersion()
