+++ platformtools/releasefetcher/src/commonMain/kotlin/io/github/kdroidfilter/platformtools/releasefetcher/github/VersionUtils.kt
@@ -1,0 +1,4 @@
+package io.github.kdroidfilter.platformtools.releasefetcher.github
+
+/** Returns the current application version. */
+expect fun getCurrentAppVersion(): String
