+++ platformtools/releasefetcher/src/wasmJsMain/kotlin/io/github/kdroidfilter/platformtools/releasefetcher/downloader/PlatformFile.wasmJs.kt
@@ -1,0 +1,3 @@
+package io.github.kdroidfilter.platformtools.releasefetcher.downloader
+
+actual class PlatformFile(val name: String, val bytes: ByteArray)
