+++ platformtools/releasefetcher/src/wasmJsMain/kotlin/io/github/kdroidfilter/platformtools/releasefetcher/github/VersionUtils.wasmJs.kt
@@ -1,0 +1,3 @@
+package io.github.kdroidfilter.platformtools.releasefetcher.github
+
+actual fun getCurrentAppVersion(): String = "0.0.0"
